+++ server/src/main/kotlin/com/papylog/plugins/Routing.kt
@@ -1,4 +1,4 @@
-package com.example.plugins
+package com.papylog.plugins
 
 import io.ktor.server.application.*
 import io.ktor.server.response.*
