+++ server/src/main/resources/application.yaml
@@ -1,6 +1,6 @@
 ktor:
     application:
         modules:
-            - com.example.ApplicationKt.module
+            - com.papylog.ApplicationKt.module
     deployment:
         port: 8080
